requirements.txt
Crawl4AI==0.4.248
faiss_cpu==1.10.0
langchain==0.3.18
langchain_community==0.3.17
langchain_core==0.3.35
langchain_ollama==0.2.3
langchain_text_splitters==0.3.6
python-dotenv==1.0.1
streamlit==1.42.0